// Frontend/HospitalManagement/src/app/model/user.model.ts
export class User {
  id?: number;
  userCode?: number;
  name?: string;
  username?: string;
  email?: string;
  password?: string;
  confirmPassword?: string;
  userRole?: string; // Backend Enum will map string automatically
  imageUrl?: string;
}










// export class User{
//     id :number|undefined;
//     userCode :number|undefined;
//     name :string|undefined;
//     username :string|undefined;
//     email :string|undefined;
//     password :string|undefined;
//     confirmPassword :string|undefined;
//     userRole :string | undefined;
//     imageUrl : string| undefined;
// }